+++ Sources/Audiograph/Audiograph.swift
@@ -21,6 +21,16 @@ public enum PlayingDuration {
     case exactly(DispatchTimeInterval)
 }
 
+/// Provides a set of smoothing options to control this step of prepreocessing before an Audiograph is played.
+public enum SmoothingOption {
+    /// Do not use any smoothing. The input data are not altered.
+    case none
+    /// Use the default smoothing setting.
+    case `default`
+    /// Provide a custom setting for computing a smoother graph. That value must be in between `[0, 1]` where `1` means the original data is used and `0` indicates maximal smoothness *(most likely a steady line)*.
+    case custom(Double)
+}
+
 /// Indicates that this view is capable of playing an Audiograph. If so, it must provide the right data to the system but stopping the playback when the view loses focus is done automatically.
 public protocol AudiographPlayable: UIView, AudiographProvider {
 }
@@ -81,6 +91,47 @@ public final class Audiograph {
     }
     
     /**
+     Before playing the chart data as Audiograph, the input can be smoothened to give the user a sound with less volatility to large spikes.
+     
+     This step is useful in cases where the user rather is interested in a trend, not in every detail of the chart.
+     The value is in between `[0, 1]` where `1` means the original data is used and `0` indicates maximal smoothness *(most likely a steady line)*.
+
+     Image it as an moving average where values in the past matter less than the more recent ones.
+     It's recommended to use the `.default` value but it can be turned off completely (`.none`) or fine-tuned to a custom value `.custom(Double)`.
+     
+     Before the Audiograph is produced, the library makes the curve smoother in order to produce an audio file that is not that volatile to large spikes. This step is useful in cases where the user rather is interested in a trend, not in every detail of the chart.
+     
+     **For example** this input graph:
+     ```
+                                    
+                               _   /
+                              / \_/
+              _   _   _     _/
+         -   / \_/ \_/ \   /
+        / \_/           \_/
+      _/
+     /
+     
+     ```
+     Will be sound more like this:
+     ```
+                             /
+                           _/
+                          /
+          ____________   /
+        _/            \_/
+       /
+      /
+     /
+     
+     ```
+     */
+    public var smoothing: SmoothingOption {
+        get { dataProcessor.smoothing }
+        set { dataProcessor.smoothing = newValue }
+    }
+    
+    /**
      This value can be used to control the final loudness volume.
      
      The final volume is computed by multiplying the default volume by that value. Default is `1.0` to apply standard loundness.
@@ -123,7 +174,7 @@ public final class Audiograph {
     /// Use a custom accessibility action retrieved from `Audiograph.createCustomAccessibilityAction(using:)` or `Audiograph.createCustomAccessibilityAction(for:)` in your view. This will play the Audiograph automatically.
     ///
     ///Audiograph can also be started by calling `Audiograph.play(graphContent:completion:)` passing in the points that are used to draw the UI.
-    /// - Parameter localizations: <#localizations description#>
+    /// - Parameter localizations: Information to fill the parts that are not providable by the library such as interaction indication phrases.
     public init(localizations: AudiographLocalizations) {
         self.localizationConfigurations = localizations
         synthesizer.completionIndicationString = localizations.completionIndicationUtterance